requirements.txt
six==1.16.0
kafka-python==2.0.2
requests==2.31.0
sseclient-py==1.8.0
